+++ src/snowflake/cli/_plugins/custom_images/commands.py
@@ -29,8 +29,10 @@ app = SnowTyperFactory(
     help="Manages custom images for Snowpark Container Services.",
 )
 
-# Required to prevent Typer from collapsing the single-subcommand group into a flat command.
-app.callback()(lambda: None)
+
+@app.callback()
+def _callback():
+    pass
 
 
 @app.command(requires_connection=False)
